move name analysis functions to separate class
